Moving `ShowAllSpecialists` into `SpecialistList` and renaming `SpecialistList` to `Specialists`.
